axis_backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=23.5
django-cors-headers>=4.3
psycopg2-binary>=2.9
msal>=1.26
celery>=5.3
redis>=5.0
python-dotenv>=1.0
gunicorn>=21.2
whitenoise>=6.6
django-extensions>=3.2
cuid2>=2.0